+++ pyrolite/extensions/__init__.py
@@ -1,0 +1,4 @@
+"""
+A stub module to house pyrolite extensions,
+where they are installed.
+"""